Code for Arduino Nano inside RC car
May not be final code, but does appear to work. Includes code to reduce likelihood of damage to front 'steering' DC motor by only turning for 600 milliseconds at full power, then reducing the power to this motor to a 'hold' level only. A video confirming the code is working can be found at https://youtu.be/m04mfECTdz8