C64 controls an RC car - code update - Replay
This Word document shows progress of my Commodore 64 project that will control an RC car using the joystick and sent via RS232 via an Arduino with a radio transceiver to an RC car. The code now allows a sequence to be stored in memory, played back, and saved to disk.