Update 21st Oct code for playing saved sequences
The code (written in 6510 assembly language) can now read the directory of a disk and display a list of any SEQ files it finds on the disk. As the sequenced saved for using the RC car are stored as SEQ files, this neatly only shows files that are relevant for loading. Once the user has selected the file they want to load, the next step is to load this into memory and offer the user to play this sequence back.