Update October 24 - All the code now working!
In an attempt to improve the bounce handling when detecting the fire button of the joystick, I was getting rather frustrated as instead of fixing this one issue, I was introducing others. I’m still not completely satisfied with the reliability of using the fire button to toggle recording on and off, but at least all the code is now 100% otherwise working well.